Added deamon option.
When specified, `daemon` argument makes it possible to run verifybam in background as a service.
This option helps with better memory usage in the long run if verifybam is used many times successively.

Most of configuration options that are given as arguments in runtime should be specified when starting
new verifybam server. These include threads count, reference genome and etc.

verifybam understands the current running daemon module and initializes itself accordingly. When there is
an already running service, verifybam expects to act as a client.